Remove duplicated code, add transform aliases and split sim into focused submodules
