+++ src/batsim/__init__.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from ._version import __version__
 
 version = __version__
@@ -10,5 +9,28 @@ try:
 except ImportError:
     _gsinterface = None
 
-from .sim import *
-from .transforms import *
+from .sim import clear_backend_memory, simulate_galaxy
+from .stamp import Stamp
+from .transforms import (
+    AffineLensingTransform,
+    FlexionTransform,
+    IATransform,
+    IaTransform,
+    LensTransform,
+)
+
+__all__ = [
+    "__version__",
+    "version",
+    "simulate_galaxy",
+    "clear_backend_memory",
+    "Stamp",
+    "LensTransform",
+    "AffineLensingTransform",
+    "IaTransform",
+    "IATransform",
+    "FlexionTransform",
+    "WCS",
+    "pltutil",
+    "stamp",
+]
